# api-specs/openapi.yaml
openapi: 3.0.3
info:
  title: Unicoach REST API
  description: REST API for the Unicoach product surface (auth, students, conversations).
  version: 1.0.0
paths:
...
  /api/v1/conversations:
    post:
      summary: Start a conversation with its first message
      description: >
        Creates a conversation owned by the caller's student and records the first
        user turn, returning the conversation plus the buffered coach reply.
        The name is derived from the first message when not supplied.
      operationId: createConversation
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: Conversation started; first turn recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateConversationResponse'
        '400':
          description: Validation failure (empty/oversized message, invalid name)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body exceeds the configured size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: The student's coaching allowance is exhausted (code coaching_budget_exhausted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The caller has no student profile (code student_profile_required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List the caller's conversations
      operationId: listConversations
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [active, archived]
            default: active
          description: active returns non-archived conversations; archived returns archived ones. Deleted conversations are never returned.
      responses:
        '200':
          description: The caller's conversations, most recent activity first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/conversations/stream:
    post:
      summary: Start a conversation and stream the first reply
      description: >
        As createConversation, but streams the coach's first reply as
        Server-Sent Events. The stream opens with a conversation event carrying the
        new conversation id and the persisted user message, then emits delta events,
        and terminates with a single message or error event.
      operationId: streamConversation
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '200':
          description: SSE stream of the first turn
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/StreamEvent'
        '400':
          description: Validation failure (rejected before the stream opens)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body exceeds the configured size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: The student's coaching allowance is exhausted (code coaching_budget_exhausted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The caller has no student profile (code student_profile_required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/conversations/{conversationId}:
    parameters:
      - in: path
        name: conversationId
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Fetch a conversation
      operationId: getConversation
      security:
        - cookieAuth: []
      responses:
        '200':
          description: The conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such conversation owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Rename and/or archive a conversation
      operationId: updateConversation
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConversationRequest'
      responses:
        '200':
          description: Updated conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '400':
          description: Validation failure (empty body, invalid name)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such conversation owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete (soft-delete) a conversation
      description: >
        Soft-deletes the conversation: it leaves all listings and can no longer be
        fetched. The underlying transcript is preserved as the memory source of
        truth. Distinct from archive, which is reversible.
      operationId: deleteConversation
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Conversation deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such conversation owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/conversations/{conversationId}/messages:
    parameters:
      - in: path
        name: conversationId
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Fetch the message history
      description: Returns all messages in the conversation in chronological order (oldest first).
      operationId: listMessages
      security:
        - cookieAuth: []
      responses:
        '200':
          description: The conversation's messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such conversation owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Post a user message and receive the coach reply
      operationId: postMessage
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMessageRequest'
      responses:
        '201':
          description: The turn was recorded; the user and coach messages are returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostMessageResponse'
        '400':
          description: Validation failure (empty/oversized message)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body exceeds the configured size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: The student's coaching allowance is exhausted (code coaching_budget_exhausted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such conversation owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: UNICOACH_SESSION
  schemas:
...
    CreateConversationRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 100000
          description: The first user message; the conversation name is derived from it when name is omitted.
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Optional caller-supplied name; server derives one from the first message when absent.
...
    UpdateConversationRequest:
      type: object
      minProperties: 1
      description: At least one field must be present. Conversations are unversioned; no version field.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        archived:
          type: boolean
          description: true archives (sets archivedAt); false unarchives (clears archivedAt).
    ConversationResponse:
      type: object
      required: [conversation]
      properties:
        conversation:
          $ref: '#/components/schemas/Conversation'
    ConversationListResponse:
      type: object
      required: [conversations]
      properties:
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
    CreateConversationResponse:
      type: object
      required: [conversation, userMessage, coachMessage]
      properties:
        conversation:
          $ref: '#/components/schemas/Conversation'
        userMessage:
          $ref: '#/components/schemas/Message'
        coachMessage:
          $ref: '#/components/schemas/Message'
...
    StreamEvent:
      oneOf:
        - $ref: '#/components/schemas/ConversationCreatedEvent'
        - $ref: '#/components/schemas/UserMessageEvent'
        - $ref: '#/components/schemas/MessageDeltaEvent'
        - $ref: '#/components/schemas/MessageCompletedEvent'
        - $ref: '#/components/schemas/StreamErrorEvent'
      discriminator:
        propertyName: type
        mapping:
          conversation: '#/components/schemas/ConversationCreatedEvent'
          user_message: '#/components/schemas/UserMessageEvent'
          delta: '#/components/schemas/MessageDeltaEvent'
          message: '#/components/schemas/MessageCompletedEvent'
          error: '#/components/schemas/StreamErrorEvent'
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          nullable: true
          description: >
            Per-field validation detail. The server emits this key explicitly as
            null when an error carries no field-level detail (e.g. 401, 409), so
            it is nullable as well as optional.
          items:
            $ref: '#/components/schemas/FieldError'